SSM secrets manager implementation in backend
