+++ catalog-rest-service/pom.xml
@@ -19,6 +19,7 @@
     <sonar.tests>${project.basedir}/src/test/java</sonar.tests>
     <sonar.tests>${project.basedir}/src/test/java</sonar.tests>
     <org.testcontainers.version>1.17.2</org.testcontainers.version>
+    <awssdk.version>2.17.243</awssdk.version>
   </properties>
 
   <dependencyManagement>
@@ -26,7 +27,7 @@
       <dependency>
         <groupId>com.amazonaws</groupId>
         <artifactId>aws-java-sdk-bom</artifactId>
-        <version>2.17.190</version>
+        <version>2.17.243</version>
         <type>pom</type>
       </dependency>
     </dependencies>
@@ -273,7 +274,12 @@
     <dependency>
       <groupId>software.amazon.awssdk</groupId>
       <artifactId>secretsmanager</artifactId>
-      <version>2.17.188</version>
+      <version>${awssdk.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>software.amazon.awssdk</groupId>
+      <artifactId>ssm</artifactId>
+      <version>${awssdk.version}</version>
     </dependency>
 
     <!--test dependencies-->
