+++ catalog-rest-service/src/main/java/org/openmetadata/catalog/secrets/AWSSecretsManager.java
@@ -15,17 +15,7 @@ package org.openmetadata.catalog.secrets;
 
 import static org.openmetadata.catalog.services.connections.metadata.OpenMetadataServerConnection.SecretsManagerProvider.AWS;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
 import com.google.common.annotations.VisibleForTesting;
-import java.util.Locale;
-import org.openmetadata.catalog.airflow.AirflowConfiguration;
-import org.openmetadata.catalog.airflow.AuthConfiguration;
-import org.openmetadata.catalog.entity.services.ServiceType;
-import org.openmetadata.catalog.exception.InvalidServiceConnectionException;
-import org.openmetadata.catalog.exception.SecretsManagerException;
-import org.openmetadata.catalog.services.connections.metadata.OpenMetadataServerConnection;
-import org.openmetadata.catalog.util.JsonUtils;
-import software.amazon.awssdk.auth.credentials.AwsBasicCredentials;
 import software.amazon.awssdk.auth.credentials.StaticCredentialsProvider;
 import software.amazon.awssdk.regions.Region;
 import software.amazon.awssdk.services.secretsmanager.SecretsManagerClient;
@@ -33,116 +23,28 @@ import software.amazon.awssdk.services.secretsmanager.model.CreateSecretRequest;
 import software.amazon.awssdk.services.secretsmanager.model.GetSecretValueRequest;
 import software.amazon.awssdk.services.secretsmanager.model.UpdateSecretRequest;
 
-public class AWSSecretsManager extends SecretsManager {
-
-  public static final String AUTH_PROVIDER_SECRET_ID_SUFFIX = "auth-provider";
-  public static final String ACCESS_KEY_ID = "accessKeyId";
-  public static final String SECRET_ACCESS_KEY = "secretAccessKey";
-  public static final String REGION = "region";
+public class AWSSecretsManager extends AWSBasedSecretsManager {
 
   private static AWSSecretsManager INSTANCE = null;
 
   private SecretsManagerClient secretsClient;
 
   private AWSSecretsManager(SecretsManagerConfiguration config, String clusterPrefix) {
-    super(AWS, clusterPrefix);
-    // initialize the secret client depending on the SecretsManagerConfiguration passed
-    if (config != null && config.getParameters() != null) {
-      String accessKeyId = config.getParameters().getOrDefault(ACCESS_KEY_ID, "");
-      String secretAccessKey = config.getParameters().getOrDefault(SECRET_ACCESS_KEY, "");
-      String region = config.getParameters().getOrDefault(REGION, "");
-      this.secretsClient =
-          SecretsManagerClient.builder()
-              .region(Region.of(region))
-              .credentialsProvider(
-                  StaticCredentialsProvider.create(AwsBasicCredentials.create(accessKeyId, secretAccessKey)))
-              .build();
-    } else {
-      // initialized with the region loaded from the DefaultAwsRegionProviderChain and credentials loaded from the
-      // DefaultCredentialsProvider
-      this.secretsClient = SecretsManagerClient.create();
-    }
+    super(AWS, config, clusterPrefix);
   }
 
   @Override
-  public Object encryptOrDecryptServiceConnectionConfig(
-      Object connectionConfig, String connectionType, String connectionName, ServiceType serviceType, boolean encrypt) {
-    String secretName =
-        buildSecretId("service", serviceType.value().toLowerCase(Locale.ROOT), connectionType, connectionName);
-    try {
-      if (encrypt) {
-        String connectionConfigJson = JsonUtils.pojoToJson(connectionConfig);
-        if (connectionConfigJson != null) {
-          upsertSecret(secretName, connectionConfigJson);
-        }
-        return null;
-      } else {
-        Class<?> clazz = createConnectionConfigClass(connectionType, extractConnectionPackageName(serviceType));
-        return JsonUtils.readValue(getSecret(secretName), clazz);
-      }
-    } catch (ClassNotFoundException ex) {
-      throw InvalidServiceConnectionException.byMessage(
-          connectionType, String.format("Failed to construct connection instance of %s", connectionType));
-    } catch (Exception e) {
-      throw SecretsManagerException.byMessage(getClass().getSimpleName(), secretName, e.getMessage());
-    }
+  void initClientWithoutCredentials() {
+    this.secretsClient = SecretsManagerClient.create();
   }
 
   @Override
-  public AirflowConfiguration encryptAirflowConnection(AirflowConfiguration airflowConfiguration) {
-    OpenMetadataServerConnection.AuthProvider authProvider =
-        OpenMetadataServerConnection.AuthProvider.fromValue(airflowConfiguration.getAuthProvider());
-    AuthConfiguration authConfig = airflowConfiguration.getAuthConfig();
-    String authProviderJson = null;
-    try {
-      switch (authProvider) {
-        case GOOGLE:
-          authProviderJson = JsonUtils.pojoToJson(authConfig.getGoogle());
-          break;
-        case AUTH_0:
-          authProviderJson = JsonUtils.pojoToJson(authConfig.getAuth0());
-          break;
-        case OKTA:
-          authProviderJson = JsonUtils.pojoToJson(authConfig.getOkta());
-          break;
-        case AZURE:
-          authProviderJson = JsonUtils.pojoToJson(authConfig.getAzure());
-          break;
-        case CUSTOM_OIDC:
-          authProviderJson = JsonUtils.pojoToJson(authConfig.getCustomOidc());
-          break;
-        case OPENMETADATA:
-          authProviderJson = JsonUtils.pojoToJson(authConfig.getOpenmetadata());
-          break;
-        case NO_AUTH:
-          break;
-        default:
-          throw new IllegalArgumentException("OpenMetadata doesn't support auth provider type " + authProvider.value());
-      }
-    } catch (JsonProcessingException e) {
-      throw new SecretsManagerException("Error parsing to JSON the auth config :" + e.getMessage());
-    }
-    if (authProviderJson != null) {
-      upsertSecret(buildSecretId(AUTH_PROVIDER_SECRET_ID_SUFFIX, authProvider.value()), authProviderJson);
-    }
-    airflowConfiguration.setAuthConfig(null);
-    return airflowConfiguration;
+  void initClientWithCredentials(String region, StaticCredentialsProvider staticCredentialsProvider) {
+    this.secretsClient =
+        SecretsManagerClient.builder().region(Region.of(region)).credentialsProvider(staticCredentialsProvider).build();
   }
 
   @Override
-  protected Object decryptAuthProviderConfig(
-      OpenMetadataServerConnection.AuthProvider authProvider, AuthConfiguration authConfig) {
-    return null;
-  }
-
-  private void upsertSecret(String secretName, String password) {
-    if (existSecret(secretName)) {
-      updateSecret(secretName, password);
-    } else {
-      storeSecret(secretName, password);
-    }
-  }
-
   public void storeSecret(String secretName, String secretValue) {
     CreateSecretRequest createSecretRequest =
         CreateSecretRequest.builder()
@@ -153,6 +55,7 @@ public class AWSSecretsManager extends SecretsManager {
     this.secretsClient.createSecret(createSecretRequest);
   }
 
+  @Override
   public void updateSecret(String secretName, String secretValue) {
     UpdateSecretRequest updateSecretRequest =
         UpdateSecretRequest.builder()
@@ -163,14 +66,7 @@ public class AWSSecretsManager extends SecretsManager {
     this.secretsClient.updateSecret(updateSecretRequest);
   }
 
-  public boolean existSecret(String secretName) {
-    try {
-      return getSecret(secretName) != null;
-    } catch (Exception e) {
-      return false;
-    }
-  }
-
+  @Override
   public String getSecret(String secretName) {
     GetSecretValueRequest getSecretValueRequest = GetSecretValueRequest.builder().secretId(secretName).build();
     return this.secretsClient.getSecretValue(getSecretValueRequest).secretString();
